requirements.txt
numpy>=1.24.0
scipy>=1.7.0
soundfile>=0.10.3
librosa>=0.10.0

# For fine-tuning
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.20.0
evaluate>=0.4.0
freesound-python-client>=0.1.0
requests-oauthlib>=1.3.1 